+++ src/main/java/com/impulsart/ImpulsArtApp/repositories/DespachoRepository.java
@@ -15,10 +15,10 @@ public interface DespachoRepository extends JpaRepository<Despacho, Long> {
     List<Despacho> findDespachos();
 
     @Query("SELECT d FROM Despacho d JOIN d.venta v JOIN d.usuario u WHERE u.identificacion = :identificacion AND d.estado != 'entregado' ORDER BY v.fechaVenta ASC")
-    List<Despacho> findDespachosAsignados(@Param("identificacion") int identificacion);
+    List<Despacho> findDespachosAsignados(@Param("identificacion") Long identificacion);
 
     @Query("SELECT d FROM Despacho d JOIN d.venta v JOIN d.usuario u WHERE u.identificacion = :identificacion AND d.estado = 'entregado' ORDER BY v.fechaVenta ASC")
-    List<Despacho> findDespachosEntregados(@Param("identificacion") int identificacion);
+    List<Despacho> findDespachosEntregados(@Param("identificacion") Long identificacion);
 
     @Query("SELECT COUNT(d) FROM Despacho d")
     int contarDespachos();
