+++ src/main/java/com/impulsart/ImpulsArtApp/service/DespachoService.java
@@ -9,10 +9,10 @@ import java.util.List;
 public interface DespachoService {
 
     public List<Despacho> findAll() throws Exception;
-    Usuario asignarDespachoAUsuario(Long pkCod_Despacho, int identificacion);
+    Usuario asignarDespachoAUsuario(Long pkCod_Despacho, Long identificacion);
     List<Despacho> findDespachos();
-    List<Despacho> findDespachosAsignados(@Param("identificacion") int identificacion);
-    List<Despacho> findDespachosEntregados(@Param("identificacion") int identificacion);
+    List<Despacho> findDespachosAsignados(@Param("identificacion") Long identificacion);
+    List<Despacho> findDespachosEntregados(@Param("identificacion") Long identificacion);
     public Despacho findById(Long pkCod_Despacho);
     public void create (Despacho despacho);
     public void update (Despacho despacho);
