+++ src/main/java/com/impulsart/ImpulsArtApp/service/OfertaService.java
@@ -9,7 +9,7 @@ public interface OfertaService {
 
     List<Oferta> findOfertasBySubasta(@Param("pkCodSubasta") Long pkCodSubasta);
     Oferta findOfertaMasAlta(@Param("pkCodSubasta") Long pkCodSubasta);
-    Oferta findBySubastaIdAndUsuarioId(Long subastaId, int usuarioId);
+    Oferta findBySubastaIdAndUsuarioId(Long subastaId, Long usuarioId);
     public List<Oferta> findAll() throws Exception;
     Oferta findById(Long PkCod_oferta);
     public void create(Oferta oferta);
