+++ src/main/java/com/impulsart/ImpulsArtApp/service/PqrsService.java
@@ -8,8 +8,8 @@ import java.util.List;
 public interface PqrsService {
 
     public List<Pqrs> findAll() throws Exception;
-    List<Pqrs> findHistorialPqrs(@Param("identificacion") Integer identificacion);
-    List<Pqrs> findPqrsAsignadoAsesores(@Param("identificacion") Integer identificacion);
+    List<Pqrs> findHistorialPqrs(@Param("identificacion") Long identificacion);
+    List<Pqrs> findPqrsAsignadoAsesores(@Param("identificacion") Long identificacion);
     public Pqrs findById(Long pkCod_Reclamo);
     public void create (Pqrs pqrs);
     public void update (Pqrs pqrs);
